% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.region.analysis.R
\name{do.region.analysis}
\alias{do.region.analysis}
\title{do.region.analysis}
\usage{
do.region.analysis(
  dat,
  sample.col,
  pop.col,
  region.col,
  area.table,
  func = "mean"
)
}
\description{
do.region.analysis
}
