% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatial.files.R
\name{read.spatial.files}
\alias{read.spatial.files}
\title{Read TIFF files into R and and create spatial data object}
\usage{
read.spatial.files()
}
\arguments{
\item{rois}{}

\item{ext}{}
}
\value{
Returns a spatial data object.
}
\description{
Read TIFF files into R and and create spatial data object
}
\examples{

}
