% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.spatial.analysis.R
\name{run.spatial.analysis}
\alias{run.spatial.analysis}
\title{run.spatial.analysis}
\usage{
run.spatial.analysis(
  dat,
  sample.col,
  pop.col,
  annot.cols = NULL,
  region.col = NULL,
  area.table = NULL,
  adj.dist = 100,
  x.col = "x",
  y.col = "y",
  distribution = TRUE,
  composition = TRUE,
  counts = TRUE,
  counts.per.area = TRUE,
  distance = TRUE,
  adjacency = TRUE,
  func = "mean"
)
}
\description{
run.spatial.analysis
}
